# backend/openapi.yaml
openapi: 3.1.0
info:
  title: WhatsApp AI Agent — Multi-Tenant SaaS API
  version: 1.0.0
  description: |
    ## Overview
    Multi-tenant restaurant ordering platform powered by an AI agent over WhatsApp.
    Each **Business** (tenant) connects a WhatsApp phone number, and the AI agent
    handles customer orders, menu inquiries, complaints, and status checks automatically.

    **Two API surfaces:**
    | Surface | Auth | Scope |
    |---------|------|-------|
    | **Tenant API** | `x-api-key` or Bearer JWT | Scoped to one business |
    | **Platform API** | Bearer JWT only (`businessId: null`) | Cross-tenant super-admin |

    ## Webhook Flow
    1. Meta sends a WhatsApp message to `POST /api/webhook`
    2. Server identifies the Business via `phone_number_id`
    3. AI agent processes the message (menu query, order submission, etc.)
    4. Reply is sent back as the response body

    ## Real-time Events (Socket.io)
    - **URL:** `http://localhost:3001`
    - **Rooms:** `business:{businessId}`
    - **Events:**
      | Event | Payload | Description |
      |-------|---------|-------------|
      | `new-message` | `{ customerId, customerPhone, message, reply }` | AI replied to a customer |
      | `new-order` | `{ orderId, referenceId, customerId, items, totalPrice }` | New order placed via WhatsApp |
      | `order-updated` | `{ orderId, status, updatedAt }` | Order status changed |
...

    ## Demo Credentials
    | Role | API Key / Endpoint | Email | Password |
...
servers:
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Business API key (e.g. `demo-api-key-123`)
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User access token from `/api/auth/login`
    PlatformBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Platform user JWT from `/api/auth/platform-login`.
        Token contains `businessId: null`.
  schemas:
...
    RestaurantSettings:
      type: object
      properties:
        name:
          type: string
        openingTime:
          type: string
          example: "09:00"
        closingTime:
          type: string
          example: "23:00"
        welcomeMsg:
          type: string
        aiRules:
          type: string
        currency:
          type: string
          example: SAR
    RestaurantSettingsInput:
      type: object
      properties:
        name:
          type: string
        openingTime:
          type: string
        closingTime:
          type: string
        welcomeMsg:
          type: string
        aiRules:
          type: string
        currency:
          type: string

    # ─── Webhook ───
...
    Role:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        isSystem:
          type: boolean
        businessId:
          type: string
          nullable: true
        permissionCount:
          type: integer
        userCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RoleInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    Permission:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
          example: orders:read
        name:
          type: string
        description:
          type: string
        category:
          type: string
          example: orders

    # ─── Audit ───
...
paths:
...
  /settings:
    get:
      tags: [Settings]
      summary: Get restaurant settings
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        '200':
          description: Restaurant settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestaurantSettings'
    put:
      tags: [Settings]
      summary: Update restaurant settings (upsert)
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestaurantSettingsInput'
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestaurantSettings'

  # ─── CONVERSATIONS ───────────────────────────────────────────────────────────
...
  /users/{id}/roles:
    post:
      tags: [Users (IAM)]
      summary: Assign a role to a user
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [roleId]
              properties:
                roleId:
                  type: string
      responses:
        '201':
          description: Role assigned
        '409':
          description: Role already assigned
  /users/{id}/roles/{roleId}:
    delete:
      tags: [Users (IAM)]
      summary: Revoke a role from a user
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: roleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Role revoked

  # ─── ROLES (IAM) ────────────────────────────────────────────────────────────
  /roles:
    get:
      tags: [Roles (IAM)]
      summary: List roles (system + business-scoped)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Roles with permission and user counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
    post:
      tags: [Roles (IAM)]
      summary: Create a custom tenant role
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleInput'
      responses:
        '201':
          description: Role created
        '409':
          description: Role name already exists
  /roles/{id}:
    get:
      tags: [Roles (IAM)]
      summary: Get role with its permissions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Role with nested permissions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Role'
                  - type: object
                    properties:
                      permissions:
                        type: array
                        items:
                          $ref: '#/components/schemas/Permission'
    put:
      tags: [Roles (IAM)]
      summary: Update role name/description
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleInput'
      responses:
        '200':
          description: Role updated
        '403':
          description: Cannot modify system roles
    delete:
      tags: [Roles (IAM)]
      summary: Delete a custom role
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Role deleted
        '403':
          description: Cannot delete system roles
  /roles/{id}/permissions:
    post:
      tags: [Roles (IAM)]
      summary: Replace all permissions on a role
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [permissionIds]
              properties:
                permissionIds:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Permissions replaced
        '403':
          description: Cannot modify system roles
  /roles/{id}/permissions/{permissionId}:
    delete:
      tags: [Roles (IAM)]
      summary: Revoke a single permission from a role
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: permissionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Permission revoked
        '403':
          description: Cannot modify system roles

  # ─── PERMISSIONS ────────────────────────────────────────────────────────────
...
  /platform/settings:
    get:
      tags: [Platform - Settings]
      summary: Get all platform settings
      security:
        - PlatformBearerAuth: []
      responses:
        '200':
          description: Platform settings as key-value map
          content:
            application/json:
              schema:
                type: object
    put:
      tags: [Platform - Settings]
      summary: Upsert platform settings
      security:
        - PlatformBearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: Key-value pairs to upsert
              example:
                default_ai_rules: "Always be polite"
                feature_whatsapp_enabled: true
      responses:
        '200':
          description: All settings after update
...